Various new features, most importantly scopes. Not tested yet....

+++ Ratty/Ratty.pm
@@ -6,7 +6,7 @@
 # Copyright (c) 2004 UK Citizens Online Democracy. All rights reserved.
 # Email: chris@mysociety.org; WWW: http://www.mysociety.org/
 #
-# $Id: Ratty.pm,v 1.11 2005-01-11 23:30:38 chris Exp $
+# $Id: Ratty.pm,v 1.12 2005-01-12 13:11:31 chris Exp $
 #
 
 package Ratty;
@@ -17,6 +17,7 @@ use DBI;
 use Digest::SHA1;
 use Error qw(:try);
 use Net::Netmask;
+use Time::HiRes;
 
 use Data::Dumper;
 
@@ -47,6 +48,7 @@ Implementation of rate-limiting.
 
 my $dbh;
 sub dbh() {
+    $dbh = undef if (defined($dbh) && defined($dbh->err()));
     $dbh ||= DBI->connect('dbi:Pg:dbname=' .  mySociety::Config::get('RATTY_DB_NAME'),
                         mySociety::Config::get('RATTY_DB_USER'),
                         mySociety::Config::get('RATTY_DB_PASS'),
@@ -58,7 +60,7 @@ sub dbh() {
 
 sub get_conditions ($) {
     my ($rule) = @_;
-    return dbh()->selectall_arrayref('select id, field, condition, value from condition where rule_id = ? order by id', {}, $rule);
+    return dbh()->selectall_arrayref('select id, field, condition, value, invert from condition where rule_id = ? order by id', {}, $rule);
 }
 
 =item new
@@ -68,16 +70,12 @@ I<Class method.> Return a new Ratty testing object.
 =cut
 sub new ($) {
     my ($class) = @_;
-    my $self = {
-            lastcommit => time(),
-            numsincelastcommit => 0,
-            lastrebuild => time(),
-            tester => compile_rules()
-        };
+    my $self = { lastrebuild => time() };
+    ($self->{lastrebuildgeneration}, $self->{tester}) = compile_rules();
     return bless($self, $class);
 }
 
-=item test VARS
+=item test SCOPE VARS
 
 I<Instance method.> Test whether the request described by VARS should be
 permitted or not. Returns true if it should not, or an array of
@@ -85,9 +83,9 @@ permitted or not. Returns true if it should not, or an array of
 
 =cut
 sub test ($$) {
-    my ($self, $V) = @_;
+    my ($self, $scope, $V) = @_;
     my $result = undef;
-    if (defined(my $r = $self->{tester}->($V))) {
+    if (defined(my $r = $self->{tester}->($scope, $V))) {
         # have a hit on rule $r
         # XXX log this
         warn "ratty rule #$r triggered\n";
@@ -98,32 +96,47 @@ sub test ($$) {
         # No rule hits, carry on.
         $result = undef;
     }
-#    ++$self->{numsincelastcommit};
-#    if ($self->{numsincelastcommit} > 50 || $self->{lastcommit} < time() - 10) {
-        dbh()->commit();
-#        $self->{numsincelastcommit} = 0;
-#        $self->{lastcommit} = time();
-#    }
-    # XXX should check for updates properly, with a trigger which updates a
-    # counter whenever the conditions or rule tables are modified.
-    if ($self->{lastrebuild} < time() - 60) {
-        $self->{tester} = compile_rules();
+
+    # Always commit. Otherwise we can deadlock, because of the way that
+    # Postgres checks foreign keys. See
+    #   http://archives.postgresql.org/pgsql-general/2004-03/msg00407.php
+    # Basically we get to pick two of,
+    #   - efficient
+    #   - concurrent
+    #   - correct
+    $dbh->commit();
+
+    if ($self->{lastrebuild} < time() - 5) {
+        my $gen = dbh()->selectrow_array('select number from generation');
+        ($self->{lastrebuildgeneration}, $self->{tester}) = compile_rules()
+            if ($gen != $self->{lastrebuildgeneration});
         $self->{lastrebuild} = time();
     }
     return $result;
 }
 
 # compile_rules
-# Return a code reference for testing requests against rate-limiting rules. The
-# returned code ref returns a rule ID when a request exceeds the rate limit, or
-# undef if it does not.
+# Return in list context the current generation number and a code reference for
+# testing requests against rate-limiting rules. The returned code ref returns a
+# rule ID when a request exceeds the rate limit, or undef if it does not; it
+# tests all scopes.
 sub compile_rules () {
-    my @code = ('sub ($$) {',
-                'my ($V, $data) = @_;',
+    my $generation = dbh()->selectrow_array('select number from generation');
+    my @code = ('sub ($$$$) {',
+                'my ($V, $scope, $data, $seen_fields) = @_;',
                 'my $dbh = Ratty::dbh();',
-                'while(my($key, $value) = each(%$V)) {',
-                '   my $num = dbh()->selectrow_array("select count(*) from available_fields where field = ?", {}, "$key");',
-                '   $dbh->do("insert into available_fields (field, example) values (?,?)", {}, "$key", "$value") if $num == 0;',
+
+                # Update the available fields table if there are fields we
+                # haven't seen before.
+                'while (my ($field, $value) = each(%$V)) {',
+                '   if (!exists($seen_fields->{$scope}) or !exists($seen_fields->{$scope}->{$field})) {',
+                '       $seen_fields->{$scope}->{$field} = 1;'
+                '       if (0 == scalar(dbh()->selectrow_array("select count(*) from available_fields where scope = ? and field = ? for update of available_fields", {}, $scope, $field))) {',
+                '           my $example = $value;'
+                '           $example = substr($value, 0, 16) . "..." if (length($eg) > 16);'
+                '           $dbh->do("insert into available_fields (scope, field, example) values (?, ?, ?)", {}, $scope, $field, $example);',
+                '       }',
+                '   }',
                 '}',
                 'my $result = undef;');
 
@@ -134,29 +147,32 @@ sub compile_rules () {
     # need to quote our input or even trust it (modulo bugs in perl).
     my @data = ();
 
-    foreach my $rule (@{dbh()->selectall_arrayref('select id, requests, interval from rule order by sequence')}) {
-        my ($ruleid, $requests, $interval) = @$rule;
+    foreach my $rule (@{dbh()->selectall_arrayref('select id, requests, interval, scope from rule order by sequence')}) {
+        my ($ruleid, $requests, $interval, $scope) = @$rule;
 
-        push(@code, 'if (1');
+        push(@data, $scope);
+        my $si = $#data;
+
+        push(@code, sprintf('if ($scope eq $data->[%d]', $si));
         my $conditions = get_conditions($ruleid);
         
         # Do local matches.
         foreach (grep { $_->[2] !~ m#[SD]# } @$conditions) {
-            my ($id, $field, $condition, $value) = @$_;
+            my ($id, $field, $condition, $value, $invert) = @$_;
             push(@data, $field);
             my $fi = $#data;
             push(@code, sprintf('&& defined($V->{$data->[%d]}) &&', $fi));
             if ($condition eq 'E') {
                 push(@data, $value);
                 my $vi = $#data;
-                push (@code, sprintf('$V->{$data->[%d]} eq $data->[%d]', $fi, $vi));
+                push (@code, sprintf('$V->{$data->[%d]} %s $data->[%d]', $fi, $invert ? 'ne' : 'eq', $vi));
             } elsif ($condition eq 'R') {
                 # Construct a regexp from the value.
                 my $re = eval(sprintf('qr#%s#', $value));
                 if (defined($re)) {
                     push(@data, $re);
                     my $vi = $#data;
-                    push(@code, sprintf('$V->{$data->[%d]} =~ m#$data->[%d]#i', $fi, $vi));
+                    push(@code, sprintf('$V->{$data->[%d]} %s m#$data->[%d]#i', $fi, $invert ? '!~' : '=~', $vi));
                 } else {
                     push(@code, '0');
                 }
@@ -165,13 +181,25 @@ sub compile_rules () {
                 if (defined($ipnet)) {
                     push(@data, $ipnet);
                     my $vi = $#data;
-                    push(@code, sprintf('$data->[%d]->match($V->{$data->[%d]})', $vi, $fi));
+                    push(@code, sprintf('%s$data->[%d]->match($V->{$data->[%d]})', $invert ? '!' : '', $vi, $fi));
                 } else {
                     push(@code, '0');
                 }
+            } elsif ($condition eq '<' or $condition eq '>') {
+                my $number = $value + 0.0;   # XXX should check that value is a number
+                if ($invert) {
+                    if ($condition eq '<') {
+                        $condition = '>=';
+                    } else {
+                        $condition = '<=';
+                    }
+                }
+                push(@data, $number);
+                my $vi = $#data;
+                push(@code, sprintf('$V->{$data->[%d]} %s $data->[%d]', $fi, $condition, $vi));
             }
         }
-        push(@code, ') {',
+        push(@code, ')) {',
                     sprintf('my ($num, $requests, $interval) = (0, %d, %d);', $requests, $interval));
 
         my @sdconds = grep { $_->[2] =~ m#^[SD]$# } @$conditions;
@@ -203,7 +231,7 @@ sub compile_rules () {
 
         # Record this hit.
         push(@code,
-            sprintf('$dbh->do(q#insert into rule_hit (rule_id, hit, shash, dhash) values (?, ?, ?, ?)#, {}, %d, time(), %s, %s);',
+            sprintf('$dbh->do(q#insert into rule_hit (rule_id, hit, shash, dhash) values (?, ?, ?, ?)#, {}, %d, Time::HiRes::time(), %s, %s);',
                     $ruleid, (@sconds ? '$S->b64digest()' : 'undef'), (@dconds ? '$D->b64digest()' : 'undef')));
 
         # Nuke old hits (XXX do this elsewhere?)
@@ -224,64 +252,78 @@ sub compile_rules () {
     #warn $codejoined;
     my $subr = eval($codejoined);
     die "evaled code: $@" if ($@);
+    
     my $D = \@data;
+    
+    # Construct hash of seen fields
+    my $S = { };
+    
+    foreach my $row (@{dbh()->selectall_arrayref('select scope, field from available_fields')}) {
+        my ($scope, $field) = @$row;
+        $S->{$scope}->{$field} = 1;
+    }
 
-    return sub ($) {
-            return &$subr($_[0], $D);
-        };
+    # Finish up the transaction.
+    dbh()->commit();
+
+    return (
+            $generation,
+            sub ($$) {
+                return &$subr($_[0], $D, $S);
+            }
+        );
 }
 
-=item admin_available_fields
+=item admin_available_fields SCOPE
 
-I<Instance method.> Returns all the fields Ratty has seen so far, and an
-example value of that field.  Structure is an array of pairs of (field,
-example).
+I<Instance method.> Returns all the fields Ratty has seen so far in the given
+SCOPE, and an example value of that field.  Structure is an array of pairs of
+(field, example).
 
 =cut
-sub admin_available_fields ($) {
-    my ($self) = @_;
-    dbh()->commit();
-    my $result = dbh()->selectall_arrayref('select field, example from available_fields');
+sub admin_available_fields ($$) {
+    my ($self, $scope) = @_;
+    my $result = dbh()->selectall_arrayref('select field, example from available_fields where scope = ?', {}, $scope);
     return $result;
 }
 
-=item admin_update_rule VALS CONDS
+=item admin_update_rule SCOPE RULE CONDITIONS
 
-I<Instance method.> Either creates a new rule or updates an existing
-one.  VALS is a hashref containing id, requests, interval, sequence and
-note (see schema.sql).  CONDS is an arrayref of conditions, each a hash
+I<Instance method.> Either creates a new rule or updates an existing one. RULE
+is a hashref containing requests, interval, sequence, note and an optional
+rule_id (see schema.sql). CONDITIONS is an arrayref of conditions, each a hash
 containing field, condition and value.
 
 =cut
-sub admin_update_rule ($$$) {
-    my ($self, $vals, $conds) = @_;
-    dbh()->commit();
+sub admin_update_rule ($$$$) {
+    my ($self, $scope, $rule, $conds) = @_;
 
     my $return = undef;
 
-    my $result = 0;
-    if ($vals->{'rule_id'}) {
-        $result = dbh()->selectrow_arrayref('select id from rule where id = ? for update', {}, $vals->{'rule_id'});
+    my $result = undef;
+    if (exists($rule->{'rule_id'})) {
+        $result = dbh()->selectrow_arrayref('select id from rule where id = ? for update', {}, $rule->{'rule_id'});
+        die "mismatch between scope \"$scope\" and rule ID \"$id\""
+            if (defined($result) and $result->{scope} ne $scope);
     }
 
-    if ($result) {
-        dbh()->do('update rule set requests = ?,
-            interval = ?, sequence = ?, note = ?, message = ? where id = ?', {}, $vals->{'requests'},
-            $vals->{'interval'}, $vals->{'sequence'}, $vals->{'note'}, $vals->{'message'},
-            $vals->{'rule_id'});
+    if (defined($result)) {
+        dbh()->do('update rule set requests = ?, interval = ?, sequence = ?, note = ?, message = ? where scope = ? and id = ?', {}, $rule->{'requests'},
+            $rule->{'interval'}, $rule->{'sequence'}, $rule->{'note'}, $rule->{'message'},
+            $scope, $rule->{'rule_id'});
     } else {
-        dbh()->do('insert into rule (requests, interval, sequence, note, message)
-            values (?, ?, ?, ?, ?)', {}, $vals->{'requests'},
-            $vals->{'interval'}, $vals->{'sequence'}, $vals->{'note'}, $vals->{'message'});
+        dbh()->do('insert into rule (scope, requests, interval, sequence, note, message)
+            values (?, ?, ?, ?, ?, ?)', {}, $scope, $rule->{'requests'},
+            $rule->{'interval'}, $rule->{'sequence'}, $rule->{'note'}, $rule->{'message'});
         $return = dbh()->selectrow_array("select currval('rule_id_seq')");
-        $vals->{'rule_id'} = $return;
+        $rule->{'rule_id'} = $return;
     }
 
     #warn(Dumper($conds));
-    dbh()->do('delete from condition where rule_id = ?', {}, $vals->{'rule_id'});
+    dbh()->do('delete from condition where rule_id = ?', {}, $rule->{'rule_id'});
     foreach my $cond (@$conds) {
         dbh()->do('insert into condition (rule_id, field, condition, value) values (?,?,?,?)',
-            {}, $vals->{'rule_id'}, $cond->{'field'},
+            {}, $rule->{'rule_id'}, $cond->{'field'},
             $cond->{'condition'}, $cond->{'value'});
     }
     
@@ -289,65 +331,74 @@ sub admin_update_rule ($$$) {
     return $return;
 }
 
-=item admin_delete_rule ID
+=item admin_delete_rule SCOPE ID
 
 I<Instance method.> Deletes the rule of the specified ID.
 
 =cut
 sub admin_delete_rule ($$$) {
-    my ($self, $id) = @_;
-    dbh()->commit();
+    my ($self, $scope, $id) = @_;
 
     my $return = undef;
 
-    dbh()->do('delete from rule_hit where rule_id = ?', {}, $id);
-    dbh()->do('delete from condition where rule_id = ?', {}, $id);
-    dbh()->do('delete from rule where id = ?', {}, $id);
-    dbh()->commit();
+    my $scope2 = dbh()->selectrow_array('select scope from rule where id = ? for update', {}, $id);
+    if (defined($scope)) {
+        if ($scope2 eq $scope) {
+            dbh()->do('delete from rule_hit where rule_id = ?', {}, $id);
+            dbh()->do('delete from condition where rule_id = ?', {}, $id);
+            dbh()->do('delete from rule where id = ?', {}, $id);
+            dbh()->commit();
+        } else {
+            die "mismatch between scope \"$scope\" and rule ID \"$id\""
+        }
+    }
 
     return $return;
 }
 
-=item admin_get_rules
+=item admin_get_rules SCOPE
 
 I<Instance method.> Returns array of hashes of data about all rules.
 
 =cut
-sub admin_get_rules ($) {
-    my ($self) = @_;
+sub admin_get_rules ($$) {
+    my ($self, $scope) = @_;
     
-    my $sth = dbh()->prepare('select * from rule order by sequence, note');
+    my $sth = dbh()->prepare('select * from rule where scope = ? order by sequence, note', {}, $scope);
     $sth->execute();
     my @ret;
-    while (my $hash_ref = $sth->fetchrow_hashref()) {
-        my $hits = dbh()->selectrow_array('select count(*) from rule_hit where rule_id = ' .  $hash_ref->{'id'});
-        $hash_ref->{'hits'} = $hits;
-        push @ret, $hash_ref;
+    while (my $rule = $sth->fetchrow_hashref()) {
+        my $hits = dbh()->selectrow_array('select count(*) from rule_hit where rule_id = ?', {}, $rule->{'id'});
+        $rule->{'hits'} = $hits;
+        push(@ret, $rule);
     }
 
     return \@ret;
 }
 
-=item admin_get_rule RULE_ID
+=item admin_get_rule SCOPE ID
 
 I<Instance method.> Returns hash of data about a rule.
 
 =cut
 sub admin_get_rule ($$) {
-    my ($self, $id) = @_;
+    my ($self, $scope, $id) = @_;
     
-    my $return = dbh()->selectall_hashref('select * from rule
-        where id = ?', 'id', {}, $id);
-    return $return->{$id};
+    my $rule = dbh()->selectrow_hashref('select * from rule where id = ?', {}, $id);
+    die "mismatch between scope \"$scope\" and rule ID \"$id\"" unless ($rule->{scope} eq $scope);
+    return $rule;
 }
  
-=item admin_get_conditions RULE_ID
+=item admin_get_conditions SCOPE ID
 
 I<Instance method.> Returns array of hashes of conditions for one rule.
 
 =cut
-sub admin_get_conditions ($$) {
-    my ($self, $id) = @_;
+sub admin_get_conditions ($$$) {
+    my ($self, $scope, $id) = @_;
+
+    my $scope2 = dbh()->selectrow_array('select scope from rule where id = ?', {}, $id);
+    die "mismatch between scope \"$scope\" and rule ID \"$id\"" unless ($scope2 eq $scope);
     
     my $sth = dbh()->prepare('select * from condition where rule_id = ?');
     $sth->execute($id);
@@ -358,4 +409,5 @@ sub admin_get_conditions ($$) {
 
     return \@ret;
 }
-  1;
+
+1;
